Moved testsmock into separate file
